Fix broken files

+++ startup-script.tpl
@@ -1,37 +1,35 @@
 #!/bin/bash
-
-source /usr/lib/bigstart/bigip-ready-functions
-
-mkdir -p  /var/log/cloud /config/cloud /var/config/rest/downloads
-
+# Setup console and startup-script logging
 LOG_FILE=/var/log/cloud/startup-script.log
-[[ ! -f $LOG_FILE ]] && touch $LOG_FILE || { echo "Run Only Once. Exiting"; exit; }
+[[ -f $LOG_FILE ]] || /usr/bin/touch $LOG_FILE
 npipe=/tmp/$$.tmp
-trap "rm -f $npipe" EXIT
-mknod $npipe p
-tee <$npipe -a $LOG_FILE /dev/ttyS0 &
+/usr/bin/trap "rm -f $npipe" EXIT
+/usr/bin/mknod $npipe p
+/usr/bin/tee <$npipe -a $LOG_FILE /dev/ttyS0 &
 exec 1>&-
 exec 1>$npipe
 exec 2>&1
 
+# skip startup script if already complete
+if [[ -f /config/startup_finished ]]; then
+  echo "Onboarding complete, skip startup script"
+  exit
+fi
 
-mkdir -p /config/cloud
+mkdir -p  /var/log/cloud /config/cloud /var/config/rest/downloads /var/lib/cloud/icontrollx_installs
 
-mkdir -p /var/lib/cloud/icontrollx_installs
 
+# Create runtime configuration on first boot
+if [[ ! -f /config/nicswap_finished ]]; then
 cat << 'EOF' > /config/cloud/runtime-init-conf.yaml
 ---
 runtime_parameters:
   - name: USER_NAME
     type: static
     value: ${bigip_username}
-  - name: SSH_KEYS
-    type: static
-    value: "${ssh_keypair}"
 EOF
 
-if ${gcp_secret_manager_authentication}
-then
+if ${gcp_secret_manager_authentication}; then
    cat << 'EOF' >> /config/cloud/runtime-init-conf.yaml
   - name: ADMIN_PASS
     type: secret
@@ -78,27 +76,19 @@ extension_services:
   service_operations:
     - extensionType: do
       type: url
-      value: https://raw.githubusercontent.com/f5devcentral/terraform-gcp-bigip-module/master/config/onboard_do.json
+      value: https://raw.githubusercontent.com/f5devcentral/terraform-gcp-bigip-module/do_issue/config/onboard_do.json
 EOF
+fi
 
-#PACKAGE_URL='https://cdn.f5.com/product/cloudsolutions/f5-bigip-runtime-init/v1.1.0/dist/f5-bigip-runtime-init-1.1.0-1.gz.run'
-for i in {1..30}; do
-curl -fv --retry 1 --connect-timeout 5 -L ${INIT_URL} -o "/var/config/rest/downloads/f5-bigip-runtime-init.gz.run" && break || sleep 10
-done
-
-
-# Install
-bash /var/config/rest/downloads/f5-bigip-runtime-init.gz.run -- '--cloud gcp'
-
-/usr/local/bin/f5-bigip-runtime-init --config-file /config/cloud/runtime-init-conf.yaml
-
+# Create nic_swap script when multi nic on first boot
 COMPUTE_BASE_URL="http://metadata.google.internal/computeMetadata/v1"
 
-if ${NIC_COUNT}
-then
+if [[ ${NIC_COUNT} && ! -f /config/nicswap_finished ]]; then
    cat << 'EOF' >> /config/cloud/nic_swap.sh
    #!/bin/bash
    source /usr/lib/bigstart/bigip-ready-functions
+   wait_bigip_ready
+   tmsh create auth user bigipuser2 password F5site02 shell bash partition-access add { all-partitions { role admin } }
    echo "before nic swapping"
    tmsh list sys db provision.1nicautoconfig
    tmsh list sys db provision.managementeth
@@ -135,12 +125,39 @@ then
    tmsh modify sys global-settings remote-host add { metadata.google.internal { hostname metadata.google.internal addr 169.254.169.254 } }
    tmsh modify sys management-dhcp sys-mgmt-dhcp-config request-options delete { ntp-servers }
    tmsh save /sys config
+   /usr/bin/touch /config/nicswap_finished
    reboot
 EOF
 fi
 
-if ${NIC_COUNT}
-then
-   chmod +x /config/cloud/nic_swap.sh
-   /config/cloud/nic_swap.sh >> $LOG_FILE
+# Create run_runtime_init.sh script on first boot
+if [[ ! -f /config/nicswap_finished ]]; then
+  cat << 'EOF' >> /config/cloud/run_runtime_init.sh
+  #!/bin/bash
+  source /usr/lib/bigstart/bigip-ready-functions
+  wait_bigip_ready
+  for i in {1..30}; do
+    curl -fv --retry 1 --connect-timeout 5 -L ${INIT_URL} -o "/var/config/rest/downloads/f5-bigip-runtime-init.gz.run" && break || sleep 10
+  done
+  bash /var/config/rest/downloads/f5-bigip-runtime-init.gz.run -- '--cloud gcp' 2>&1
+  /usr/local/bin/f5-bigip-runtime-init --config-file /config/cloud/runtime-init-conf.yaml 2>&1
+  /usr/bin/touch /config/startup_finished
+EOF
 fi
+
+# Run scripts based on number of nics
+if ${NIC_COUNT}; then
+  if [[ -f /config/nicswap_finished ]]; then
+    echo "Running run_runtime_init.sh"
+    chmod +x /config/cloud/run_runtime_init.sh
+    nohup /config/cloud/run_runtime_init.sh &
+  else
+    chmod +x /config/cloud/nic_swap.sh
+    nohup /config/cloud/nic_swap.sh &
+  fi
+else
+    echo "Running run_runtime_init.sh"
+    chmod +x /config/cloud/run_runtime_init.sh
+    nohup /config/cloud/run_runtime_init.sh &
+fi
+
